add -p/--pid option to read a process' environment
Rather than --pid taking a value, this is implemented as --pid changing
the behavior of the FILE argument. That's the only way I can think of to
make clap happy for when I add filtering patterns later, and the UX is
basically the same besides some slightly more confusing help text.